Alter how configuration information for scheduled tasks get set
Beforehand, changing the id of a task or deleting the task would
cause an exception when starting the app. Additionally, if the app
is being frequently restarted, then some of the tasks would never run,
since the apschedule module was resetting the next run time on each
new startup. The app module was given the global variable MAIN_PROCESS
to keep the spawned processes from running the same initialization.