Use jobs DB as semaphore for job tasks
Beforehand, there were instances of more than one process starting
a task at the same time. Additionally, move the jobs print statement
to a regular running task, and have it monitor the other tasks for
issues, such as a semaphore that should not be locked, or a task
starting within a certain time period of another task.