Add ability to reschedule tasks that are missed due to sleep/hibernate
This needs to be done in the main task and not in a spawned process,
so the task monitor was refactored. This requires shadowing apscheduler
ad storing the latest next run times, and then rescheduling as necessary
after a sleep/hibernate event is detected.

Additionally, a lot of the initialization code was moved to its own module
along with the reschedule code, since there was a lot of code which was
being duplicated, plus to be able to define the code using existing code
standards before the actual code gets run so that it can passed along
to the function decorators.