Add exception wrap around database session commits
An exception encountered during the process of creating/updating records
was leaving the session in an inactive status for that process. The
only way to be able to access the database again is to issue an explicit
rollback.

Only the create/update session commits were wrapped at this time. Wrapping
other session commits need to be evaluated for whether they are required
or not.