Add error handling for `<img>` elements
This was primarily done since images from other sites was now being
shown with the subscription elements, and very often those images are
dead and fail to load. However, when using the media controller that
comes with Prebooru instead of an external one like Nginx, it was also
noticed that the image would sometimes fail to load on the first try.