Add jobs info table to table creation
It was noticed that trying to create a new Prebooru instance was
failing to startup because this table was not yet created. Several
checks and updates which were added since jobs were first integrated
took advantage of the fact that the jobs table already existed, so this
issue wasn't noticed until just recently.