Fix scripts not processing all missing files
Since it uses negation in the query, it has to use the limit paginate
method. A "total" attribute was added to all paginate extensions, since
the main paginate also has this. The "total" on the limit paginate is
static, since the amount will change after every iteration, but the total
pages being processed would remain constant.