Fix rescheduling vacuum task
It can only be rescheduled from the parent process, since that is the
process that has access to the scheduler object.

Also decrease the jitter, since it will retry only every hour, which
gives it less of a chance since there has to be at least 5 minutes
since the last scheduled task has completed.