Make video samples optional
This is primarily for subscription posts, since unlike the previews,
it's very unlikely that the video sample will ever be used, and creating
the sample consumes a lot of CPU/memory resources.

Eventually though a scheduled task will need to be added that adds
video samples back onto these posts, though probably only for those
which are no longer linked to a subscription element, i.e. those that
were unlinked due to the "keep" field being set to "yes".