Drop requery from artists and illusts
The API data cache already fills this purpose. Although, a check was
put in to keep them from being updated twice within a day if something
was changed. This would not prevent it from trying multiple times though
if nothing gets changed, but this is the tradeoff for removing those
values which otherwise serve little purpose.