Add a check to ensure that the server is listening on the port
Sometimes the process would get orphaned from the console environment,
which would leave a server up but unresponsive to any changes in the
code. Any attempts to start a new server would also fail to listen on
the port. This check is a simple way to make sure that the port is active
for the current server process.