Explicitly close the database session in all opened processes/threads
This was causing hundreds of file handles to the database to be left
open, since the sessions are maintained globally by the SESSION variable
even if the thread is closed. The only thread which seem to have session
closing implicitly built in is the route handlers, so those did not
require any modification.

Along with adding this, several refactors were done to make doing the
common steps of task startup and shutdown portions automatically. Additionally,
refactoring was done to only pass basic types, i.e. integers/strings, to
threads and processes, to avoid leaving a model in a bad state when the
session gets shut down.