Allow server checks to shutdown the server process
Since this validation does not take place in the main thread, several
additional mechanisms were put in to allow for the child process to
exit the process while setting an error code.

Additionally, the returncode was never being populated because the
poll method needs to be called first, which populates that value.