Convert other tables to use the without rowid definition
No migration for these, since changing the rowid status does not affect
the operation of SQLAlchemy or the table in any way visible to the program
or the user. The same could be said of the previous commit, however these
are not part of the main schemea.

It's easy enough to switch these tables over though, such as using easy
enough using the SQLite Browser and is left as an exercise to the user.
They can also just be dropped entirely, since the program will recreate
them during the startup initialization.