Change how caching is done on class properties
The base class of the polymorphic models was overriding all of the
cached properties of it's subclasses, so instead just create a unique
name and set that as a property. This does add some additional lookup
delay, however there doesn't appear to be a way to set an attribute on
a super class that doesn't immediately propagate to subclasses.