Fix reinstantiating a post failing with no message
If the Archive records function failed for any reason, then the Subscription
records function was not returning that cause of failure. Now, in addition
to doing that, it also updates the status on the subscription if there is a
another element with the same MD5 that is unlinked, and sets the status
to duplicate if so.