Update how the illust url relation gets handled on uploads
This was causing an exception on the Upload SHOW page, since it was
referencing a site_id, which the models no longer have. Additionally,
it was still working under the assumption that the illust is only
available on network uploads through the posts model, and now it has
a more direct link through the upload elements model.