Add a scheduled task to recalculate the pool element count and positions
This logic was updated on 166e7fa to remove doing a full calculate
on ever pool element insertion/deletion, which has the potential to
leave holes in a pool. Plus if it does leave a hole, then the positions
at the end could be outside the pool element count range.

So pool element insertions will always add to the element count, since
those are always added to the end, and pool element deletions will only
subtract from the element count if they are the last element. This will
still leave holes, but at least the elements at the end won't fall
outside the boundary of the element count.

Additionally, a scheduled task was added in to do a regular recalculate
on all recently updated pools, which is now trackable through adding
a checked field which gets updated on everytime the full calculation
is done.