Add a query property method instead of a value to secondary tables
If it is just assigned as a value, then the DB session it uses will
be different from the request/processes DB session, which will then
cause the connection to be left open until the process is terminated.
Getting the main SESSION value each time when generating a query should
prevent this from happening.