Normalize hierarchy of found tweets for get_twitter_illust_timeline
Sometimes the response has the tweet info one layer deeper than what
is found, which was causing it to report that no tweets were being
found, causing it to often fallback to the get_twitter_illust function.

Also have the affected function output the actual data response upon
error, since this won't get saved to the API storage, so there needs
to be some way to debug issues.