Update check constraint on pool_element null check
The previous version allowed for one or more foreign keys to be set,
when really it should be one and only one.

Note: When dropping the constraint, an extra ck_pool_element had to
be added to the beginning of the constraint name since it seems that
alembic adds that extra bit when reflecting the table for unknown
reasons.