Switch notation to use foreign keys instead of secondary tables
This saves the creation of secondary tables, plus since the link is
more direct now, the access time should be faster. Since the pool
attachment is still using a quasi secondary table, a boolean was added
to indicate whether there was a pool attached or not.

Two check constraints were added as well:

1. A boolean check constraint, since the other check constraint relies
   on this value having a 1 or 0 value, and booleans can technically
   hold any numerical value.

2. A validation that the number of attachments can be either zero or one.