Update several Twitter graphql endpoints to use HTTP 3.0 with `httpx`
This was required as Twitter would return blank results otherwise.
The parameters for the graphql endpoints was also updated. Besides
that, the API data for a single post is now retrieved with graphql
instead of v1.1 of the API, which no longer appears to work.

For errors, check for a reason with a default of None, since `httpx`
requests do not return a reason upon error.