Add a fallback to the upload process
It was experienced that when a large number of uploads were queued,
that a certain number would not get processed and stay in the pending
state. This starts a new thread (if not already started) when an upload
is queued, and keeps checking until there are no more uploads pending,
requeuing an upload if needed.