Revise the upload all/select duplicate mechanics.
Show the upload instead of flashing an error message, and add duplicate
detection to the upload select interface.

This makes it easier to navigate to the upload and ensure that it has
all of the desired downloads. Since upload select creates media files,
it was given a mechanism to force the loading of the image previews.