Move artist update function out of records package
This particular version of the update keeps information from being
removed, such as names and site accounts, which should only occur
when the user does them.

By doing this, this also removes all records package imports from the
sources package, which allows the removal of any function imports
because of import conflicts.