Add a fix script for removing duplicate values in the M2M tables
It's possible for there to be duplicates since the text fields of the
label and description table are not unique. Two threads attempting to
add the same value would query and find that it doesn't exist, then
add the same value not knowing that another thread was doing the same.

This fix script is designed to only be run once, and to be run before
all of the major migrations that occur in 2.36, so that when those
migrations do occur, that the values in those two text tables are
truly unique.