Add MD5 column to illust URL
This is in preparation to removing the MD5 columns from the download
elements and subscription elements. The linking between illust URLs
and posts will also be changed to using the MD5. This change will also
allow illust URLs to be linked to archive posts. This will make getting
the archives for subscription elements significantly simpler.