Make md5 the primary link between illust URLs and posts
This allows the post_id to be removed from illust URLs, and illusts
to be removed from archive posts. This automatically relinks posts
and illust URLs when either is reintantiated.

Since neither the post md5s nor the illust URL md5s are guaranteed
to be subsets of each other, no foreign keys can be added to the
database schema, which is why the relationships have to be defined
explicitly in the model files.

As a side benefit of all this, archive posts are now linked to illust
URLs, which allows them to be easily used in places such as on
subscription elements.