Fix issue with batch loader
The batch loader assumes that the relation will be set up with the
primary join going a certain direction, which is not always the case.

Additionally, this was discvored on the download SHOW page, which
should not be using the batch loader and instead the selectin loader,
so selectin functions were added for the illust and artist as well.